requirements.txt
langgraph
arxiv
pydantic
serpapi
google-search-results
python-dotenv